adds custom hook
