+++ src/hooks/use-scripts.js
@@ -1,0 +1,20 @@
+// copied from stackoverflow, do not trust
+// https://stackoverflow.com/questions/34424845/adding-script-tag-to-react-jsx
+import { useEffect } from "react"
+
+const useScript = (url) => {
+  useEffect(() => {
+    const script = document.createElement("script")
+
+    script.src = url
+    script.async = true
+
+    document.body.appendChild(script)
+
+    return () => {
+      document.body.removeChild(script)
+    }
+  }, [url])
+}
+
+export default useScript
